Get basic API requests with auth working
The way this works is kind of convoluted. To make API requests to
GitHub, you create a JWT signed with your private key (fine). That
doesn't work to make API requests to a repository we're installed on, so
what we actually do is:

1. grab our installation id from the webhook POST
2. post to /app/installations/{installation_id}/access_tokens (using JWT
   auth), which will give you a different token, valid for 1h
3. make api calls against the repository with _that_ token.

This code still doesn't do anything useful, but now it _could_, if it
were so inclined.